pyproject.toml
[tool.poetry]
name = "simple-crud"
version = "0.1.0"
description = ""
authors = ["George Zografos <george.p.zografos@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
tortoise-orm = "^0.20.0"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
asyncpg = "^0.28.0"
python-dotenv = "^1.0.0"
aerich = "^0.7.2"
httpx = "^0.25.0"
passlib = "^1.7.4"
python-multipart = "^0.0.6"
bcrypt = "^4.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.10.1"
mypy = "^1.6.1"
isort = "^5.12.0"
flake8 = "^6.1.0"

[tool.aerich]
tortoise_orm = "backend.settings.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
